Crate wasmer_artifact

Generic Artifact abstraction for Wasmer Engines.

Modules

artifact

error

The WebAssembly possible errors

funcbody

Structs

FunctionBodyPtr

A safe wrapper around VMFunctionBody.

MetadataHeader

Metadata header which holds an ABI version and the length of the remaining metadata.

VMFunctionBody

A placeholder byte-sized type which is just used to provide some amount of type safety when dealing with pointers to JIT-compiled function bodies. Note that it’s deliberately not Copy, as we shouldn’t be carelessly copying function body bytes around.

Enums

DeserializeError

The Deserialize error can occur when loading a compiled Module from a binary.

ImportError

An ImportError.

PreInstantiationError

An error while preinstantiating a module.

SerializeError

The Serialize error can occur when serializing a compiled Module into a binary.

Constants

VERSION

Version number of this crate.

Traits

ArtifactCreate

An Artifact is the product that the Engine implementation produce and use.

Upcastable

Trait needed to get downcasting of Engines to work.