Struct wasmer_artifact::MetadataHeader

#[repr(C)]
pub struct MetadataHeader {
    magic: [u8; 8],
    version: u32,
    len: u32,
}

Metadata header which holds an ABI version and the length of the remaining metadata.

Fields
magic: [u8; 8]version: u32len: u32
Implementations

impl MetadataHeader

const CURRENT_VERSION: u32

Current ABI version. Increment this any time breaking changes are made to the format of the serialized data.

const MAGIC: [u8; 8]

Magic number to identify wasmer metadata.

pub const LEN: usize

Length of the metadata header.

pub const ALIGN: usize

Alignment of the metadata.

pub fn new(len: usize) -> [u8; 16]

Creates a new header for metadata of the given length.

pub fn parse(bytes: &[u8]) -> Result<usize, DeserializeError>

Parses the header and returns the length of the metadata following it.